refactor: use only uint32_t and int32_t in the widget's API

+++ src/core/lv_global.h
@@ -74,7 +74,7 @@ typedef struct _lv_global_t {
     lv_ll_t style_trans_ll;
     bool style_refresh;
     uint32_t style_custom_table_size;
-    uint16_t style_last_custom_prop_id;
+    uint32_t style_last_custom_prop_id;
     uint8_t * style_custom_prop_flag_lookup_table;
 
     lv_ll_t group_ll;
@@ -185,7 +185,7 @@ typedef struct _lv_global_t {
 #endif
 
 #if LV_USE_IME_PINYIN != 0
-    uint16_t ime_cand_len;
+    size_t ime_cand_len;
 #endif
 } lv_global_t;
