+++ src/widgets/animimage/lv_animimage.c
@@ -64,13 +64,13 @@ lv_obj_t * lv_animimg_create(lv_obj_t * parent)
     return obj;
 }
 
-void lv_animimg_set_src(lv_obj_t * obj, const void * dsc[], uint8_t num)
+void lv_animimg_set_src(lv_obj_t * obj, const void * dsc[], size_t num)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
     lv_animimg_t * animimg = (lv_animimg_t *)obj;
     animimg->dsc = dsc;
     animimg->pic_count = num;
-    lv_anim_set_values(&animimg->anim, 0, num - 1);
+    lv_anim_set_values(&animimg->anim, 0, (int32_t)num - 1);
 }
 
 void lv_animimg_start(lv_obj_t * obj)
@@ -92,7 +92,7 @@ void lv_animimg_set_duration(lv_obj_t * obj, uint32_t duration)
     lv_anim_set_playback_delay(&animimg->anim, duration);
 }
 
-void lv_animimg_set_repeat_count(lv_obj_t * obj, uint16_t count)
+void lv_animimg_set_repeat_count(lv_obj_t * obj, uint32_t count)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
     lv_animimg_t * animimg = (lv_animimg_t *)obj;
@@ -124,7 +124,7 @@ uint32_t lv_animimg_get_duration(lv_obj_t * obj)
     return lv_anim_get_time(&animimg->anim);
 }
 
-uint16_t lv_animimg_get_repeat_count(lv_obj_t * obj)
+uint32_t lv_animimg_get_repeat_count(lv_obj_t * obj)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
     lv_animimg_t * animimg = (lv_animimg_t *)obj;
